Validation for updating articles

+++ src/main/java/com/patternson/webshopclient/model/ArticleDTO.java
@@ -1,19 +1,15 @@
 package com.patternson.webshopclient.model;
 
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.hibernate.validator.constraints.NotBlank;
-import org.hibernate.validator.constraints.NotEmpty;
 
 import javax.validation.constraints.Max;
 import javax.validation.constraints.Min;
-import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
 import java.math.BigDecimal;
-import java.math.BigInteger;
 
 /**
  *
@@ -44,7 +40,7 @@ public class  ArticleDTO {
 
     @Min(0)
     @Max(9999)
-    private BigInteger stock;
+    private Integer stock;
 
     @JsonProperty("article_url")
     private String articleUrl;
